gerenciamemos/requirements-dev.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-tinymce==2.6.0
python-decouple==3.1
pytz==2017.2
static3==0.7.0
